ready

+++ impl/src/main/java/ru/skillbox/diplom/group40/social/network/impl/service/passRecovery/RecoveryService.java
@@ -1,12 +1,17 @@
 package ru.skillbox.diplom.group40.social.network.impl.service.passRecovery;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.skillbox.diplom.group40.social.network.api.dto.auth.NewPasswordDto;
 import ru.skillbox.diplom.group40.social.network.domain.account.Account;
 import ru.skillbox.diplom.group40.social.network.domain.passwordRecovery.RecoveryToken;
+import ru.skillbox.diplom.group40.social.network.domain.user.User;
+import ru.skillbox.diplom.group40.social.network.impl.exception.AuthException;
 import ru.skillbox.diplom.group40.social.network.impl.mapper.account.MapperAccount;
 import ru.skillbox.diplom.group40.social.network.impl.repository.recoveryToken.RecoveryTokenRepository;
+import ru.skillbox.diplom.group40.social.network.impl.repository.user.UserRepository;
 import ru.skillbox.diplom.group40.social.network.impl.service.account.AccountService;
 import ru.skillbox.diplom.group40.social.network.impl.utils.mail.MailUtil;
 
@@ -16,23 +21,25 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@Transactional
 @RequiredArgsConstructor
 public class RecoveryService {
-    private final AccountService accountService;
+    private final UserRepository userRepository;
     private final RecoveryTokenRepository tokenRepository;
     private final MailUtil mailUtil;
-    private final MapperAccount accountMapper;
-    private final String host = "localhost:8099";
+    @Value("${spring.mail.linkHost}")
+    private String host;
 
-    public Boolean recover(String email) {
-        Account account = accountService.getAccountByEmail(email);
-        if (account == null) {
-            return false;
+    public void recover(String email) {
+        Optional<User> optionalUser = userRepository.findFirstByEmail(email);
+        if (optionalUser.isEmpty()) {
+            throw new AuthException("no such user");
         }
-        RecoveryToken token = new RecoveryToken(account, LocalDateTime.now().plus(10, ChronoUnit.MINUTES));
+        RecoveryToken token = new RecoveryToken(
+                optionalUser.get().getId(),
+                LocalDateTime.now().plus(10, ChronoUnit.MINUTES));
         tokenRepository.save(token);
         mailUtil.send(email, getMessage(token.getId()));
-        return true;
     }
 
     private String getMessage(UUID tokenId) {
@@ -40,19 +47,22 @@ public class RecoveryService {
                 "http://" + host + "/change-password/" + tokenId.toString();
     }
 
-    public boolean setNewPassword(String linkId, NewPasswordDto passwordDto) {
+    public void setNewPassword(String linkId, NewPasswordDto passwordDto) {
         Optional<RecoveryToken> optionalToken = tokenRepository.findById(UUID.fromString(linkId));
         if (optionalToken.isEmpty()) {
-            return false;
+            throw new AuthException("wrong link");
         }
         RecoveryToken token = optionalToken.get();
         if (token.getExpirationTime().isBefore(LocalDateTime.now())) {
-            return false;
+            throw new AuthException("recovery token expired");
         }
-        Account account = token.getAccount();
-        account.setPassword(passwordDto.getPassword());
-        accountService.save(account);
+        Optional<User> optionalUser = userRepository.findFirstById(token.getUserId());
+        if(optionalUser.isEmpty()){
+            throw new AuthException("no such user");
+        }
+        User user = optionalUser.get();
+        user.setPassword(passwordDto.getPassword());
+        userRepository.save(user);
         tokenRepository.delete(token);
-        return true;
     }
 }
